Extract doctest::IsNaN to doctest/parts/matchers/is_nan.{h,cpp}
