+++ doctest/parts/doctest_fwd.h
@@ -67,6 +67,7 @@ DOCTEST_MSVC_SUPPRESS_WARNING(4623) // default constructor was implicitly define
 #include <doctest/parts/public/string.h>
 #include <doctest/parts/public/matchers/contains.h>
 #include <doctest/parts/public/matchers/approx.h>
+#include <doctest/parts/public/matchers/is_nan.h>
 
 namespace doctest {
 
@@ -325,23 +326,6 @@ struct ContextOptions //!OCLINT too many fields
 };
 
 DOCTEST_INTERFACE const ContextOptions* getContextOptions();
-
-template <typename F>
-struct DOCTEST_INTERFACE_DECL IsNaN
-{
-    F value; bool flipped;
-    IsNaN(F f, bool flip = false) : value(f), flipped(flip) { }
-    IsNaN<F> operator!() const { return { value, !flipped }; }
-    operator bool() const;
-};
-#ifndef __MINGW32__
-extern template struct DOCTEST_INTERFACE_DECL IsNaN<float>;
-extern template struct DOCTEST_INTERFACE_DECL IsNaN<double>;
-extern template struct DOCTEST_INTERFACE_DECL IsNaN<long double>;
-#endif
-DOCTEST_INTERFACE String toString(IsNaN<float> in);
-DOCTEST_INTERFACE String toString(IsNaN<double> in);
-DOCTEST_INTERFACE String toString(IsNaN<double long> in);
 
 #ifndef DOCTEST_CONFIG_DISABLE
